Update branch matching and caching in GHA for #19
My initial setup was naive, but we now need to process revisions and the
branches separately for keys and restoreKeys of hashes. This also means
acquiring the branch names and their current HEAD commits dynamically as
the format will not be known a-priori and their contents will be more
unstable than tagged revisions.